pyproject.toml
[project]
name = "mcp-getting-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.31.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
]
